# docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PairingClaimIn": {
        "properties": {
          "device_uid": {
            "maxLength": 128,
            "minLength": 4,
            "title": "Device Uid",
            "type": "string"
          },
          "pairing_code": {
            "maxLength": 32,
            "minLength": 4,
            "title": "Pairing Code",
            "type": "string"
          }
        },
        "required": [
          "device_uid",
          "pairing_code"
        ],
        "title": "PairingClaimIn",
        "type": "object"
      },
      "PairingClaimOut": {
        "properties": {
          "claimed_at": {
            "format": "date-time",
            "title": "Claimed At",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "device_token": {
            "title": "Device Token",
            "type": "string"
          },
          "device_uid": {
            "title": "Device Uid",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "integer"
          }
        },
        "required": [
          "device_id",
          "owner_user_id",
          "device_uid",
          "device_token",
          "claimed_at"
        ],
        "title": "PairingClaimOut",
        "type": "object"
      },
      "PairingStartIn": {
        "properties": {
          "device_uid": {
            "maxLength": 128,
            "minLength": 4,
            "title": "Device Uid",
            "type": "string"
          }
        },
        "required": [
          "device_uid"
        ],
        "title": "PairingStartIn",
        "type": "object"
      },
      "PairingStartOut": {
        "properties": {
          "device_uid": {
            "title": "Device Uid",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "pairing_code": {
            "title": "Pairing Code",
            "type": "string"
          },
          "ttl_seconds": {
            "title": "Ttl Seconds",
            "type": "integer"
          }
        },
        "required": [
          "device_uid",
          "pairing_code",
          "expires_at",
          "ttl_seconds"
        ],
        "title": "PairingStartOut",
        "type": "object"
      },
      "PairingStatusOut": {
        "properties": {
          "claimed": {
            "title": "Claimed",
            "type": "boolean"
          },
          "claimed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "device_uid": {
            "title": "Device Uid",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "pairing_code": {
            "title": "Pairing Code",
            "type": "string"
          },
          "ttl_seconds": {
            "title": "Ttl Seconds",
            "type": "integer"
          }
        },
        "required": [
          "pairing_code",
          "device_uid",
          "claimed",
          "expires_at",
          "ttl_seconds"
        ],
        "title": "PairingStatusOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Device-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/devices/pairing/confirm": {
      "post": {
        "description": "Device -> Backend (ohne JWT):\n- validiert pairing_code (existiert, nicht used, nicht expired)\n- setzt device.owner_user_id + is_claimed\n- erstellt DeviceToken (plaintext nur einmal zur\u00fcckgeben)",
        "operationId": "confirm_pairing_api_v1_devices_pairing_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingClaimOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Confirm Pairing",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/devices/pairing/start": {
      "post": {
        "description": "User -> Backend:\n- stellt sicher, dass das Device existiert\n- erstellt eine PairingSession mit Ablaufzeit",
        "operationId": "start_pairing_api_v1_devices_pairing_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingStartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Pairing",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/devices/pairing/status": {
      "get": {
        "description": "Read-only status for a pairing_code.\n- claimed=true when code has been used/claimed by a user",
        "operationId": "pairing_status_api_v1_devices_pairing_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "pairingCode",
            "required": true,
            "schema": {
              "title": "Pairingcode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Pairing Status",
        "tags": [
          "pairing"
        ]
      }
    },
...
    "/api/v1/pairing/confirm": {
      "post": {
        "description": "Device -> Backend (ohne JWT):\n- validiert pairing_code (existiert, nicht used, nicht expired)\n- setzt device.owner_user_id + is_claimed\n- erstellt DeviceToken (plaintext nur einmal zur\u00fcckgeben)",
        "operationId": "confirm_pairing_api_v1_pairing_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingClaimOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Confirm Pairing",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/pairing/start": {
      "post": {
        "description": "User -> Backend:\n- stellt sicher, dass das Device existiert\n- erstellt eine PairingSession mit Ablaufzeit",
        "operationId": "start_pairing_api_v1_pairing_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingStartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Start Pairing",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/pairing/status": {
      "get": {
        "description": "Read-only status for a pairing_code.\n- claimed=true when code has been used/claimed by a user",
        "operationId": "pairing_status_api_v1_pairing_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "pairingCode",
            "required": true,
            "schema": {
              "title": "Pairingcode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Pairing Status",
        "tags": [
          "pairing"
        ]
      }
    },
...
  }
}